Double precision

+++ src/main/java/com/cluttered/cryptocurrency/ann/NeuralNetwork.java
@@ -4,7 +4,6 @@ package com.cluttered.cryptocurrency.ann;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.math.BigDecimal;
 import java.util.List;
 
 /**
@@ -14,22 +13,22 @@ public class NeuralNetwork {
 
     private static final Logger LOG = LoggerFactory.getLogger(NeuralNetwork.class);
 
-    private final Long inputSize;
+    private final Integer inputSize;
     private final List<Layer> layers;
 
-    public NeuralNetwork(final Long inputSize, final List<Layer> layers) {
+    public NeuralNetwork(final Integer inputSize, final List<Layer> layers) {
         this.inputSize = inputSize;
         this.layers = layers;
     }
 
-    public List<BigDecimal> fire(final List<BigDecimal> inputs) {
+    public List<Double> fire(final List<Double> inputs) {
         final long startTimeMillis = System.currentTimeMillis();
         LOG.info("########## Fire NeuralNetwork ##########");
         LOG.info("Inputs: {}", inputs);
         if (inputs.size() != inputSize)
             throw new IllegalArgumentException("NeuralNetwork accepts " + inputSize + " inputs but received " + inputs.size());
 
-        List<BigDecimal> layerResults = inputs;
+        List<Double> layerResults = inputs;
         for (final Layer layer : layers) {
             layerResults = layer.fire(layerResults);
         }
